split what was in applicaton helper into separate files
